msm: camera: Fix the stretch issue in Camera Preview
In non zsl camera preview mode, two vfe outputs are configured as
preview and video outputs. Greater of the two is output from main
scaler and the lower one from the secondary scaler. In Camera preview
mode,
application matches preview dimensions aspect ratio to snapshot
dimensions and in camcorder mode it is matched to video.

In Current design, We have below scenarios for Non ZSL Camera Preview

Case1: Main Scaler is preview and secondary scalar is video
In this scenario no error is observed
Case2: Main scalar is video and secondary scalar is preview
In this scenario since preview output in derived from video but the
aspect ratio may not match in camera mode, so we see stretch in camera
preview.

This issue is fixed by configuring only single VFE output as per
display dimensions in Camera Preview  since the second video output
is never used

CRs-Fixed: 343884
Change-Id: I25a82b417a3ed1b28ea243c6ce988c23d5fed4a7
Signed-off-by: Nishant Pandit <npandit@codeaurora.org>